Workaround spice-protocol 0.12.8 regression
spice-protocol 0.12.8 in combination with spice-server 0.12.5 or older
causes compilation breakage. This does not happen upstream as
spice-server code was changed to avoid that. In particular, this bug
causes QEMU to be built without SPICE support.

A new spice-protocol release will be needed to fix that regression, in
the mean time I backported the upstream spice-server patch which should
resolve the compilation issues for spice-server users.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1239102